Enforce upload size limits and refactor service location
