+++ app/common/src/main/java/stirling/software/common/exception/FileTooLargeException.java
@@ -1,0 +1,9 @@
+package stirling.software.common.exception;
+
+/** Exception thrown when an uploaded file exceeds the configured maximum size. */
+public class FileTooLargeException extends RuntimeException {
+
+    public FileTooLargeException(long maxSize) {
+        super("File exceeds maximum size of " + maxSize + " bytes");
+    }
+}
