+++ app/core/src/main/java/stirling/software/SPDF/model/api/HandleDataRequest.java
@@ -7,12 +7,23 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import stirling.software.common.util.GeneralUtils;
+
 @Data
 @EqualsAndHashCode
 public class HandleDataRequest {
 
     @Schema(description = "The input files", requiredMode = Schema.RequiredMode.REQUIRED)
     private MultipartFile[] fileInput;
+
+    public MultipartFile[] getFileInput() {
+        if (fileInput != null) {
+            for (MultipartFile file : fileInput) {
+                GeneralUtils.checkMaxUploadSize(file);
+            }
+        }
+        return fileInput;
+    }
 
     @Schema(
             description = "JSON String",
