+++ app/core/src/main/java/stirling/software/SPDF/model/api/MultiplePDFFiles.java
@@ -7,9 +7,20 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import stirling.software.common.util.GeneralUtils;
+
 @Data
 @EqualsAndHashCode
 public class MultiplePDFFiles {
     @Schema(description = "The input PDF files", requiredMode = Schema.RequiredMode.REQUIRED)
     private MultipartFile[] fileInput;
+
+    public MultipartFile[] getFileInput() {
+        if (fileInput != null) {
+            for (MultipartFile file : fileInput) {
+                GeneralUtils.checkMaxUploadSize(file);
+            }
+        }
+        return fileInput;
+    }
 }
